Revert "cleanup"
This reverts commit 28ccbf0d63f95f7fdb8797a9bed712f049d4bf49.

+++ SbankenClient/Models/Account.swift
@@ -14,8 +14,8 @@ public class Account: Codable {
     public let ownerCustomerId: String
     public let name: String
     public let accountType: String
-    public let available: Float
-    public let balance: Float
-    public let creditLimit: Float
+    public let available: Double
+    public let balance: Double
+    public let creditLimit: Double
     public let defaultAccount: Bool
 }
